+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-01T08:41:10.891Z",
+    "2026-07-01T10:48:33.211Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {