+++ packages/orq-rc/RELEASES.md
@@ -10419,3 +10419,13 @@ Based on:
 - [typescript v4.11.0-rc.64] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.64] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.64 - packages/orq-rc
+
+## 2026-07-01 10:51:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.65] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.65] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.65 - packages/orq-rc